fix: cast k_trans/v to float32 in _fwd_kv_parallel Triton kernel
tl.dot requires both operands to be same dtype. k_decay is fp32
so k_trans*k_decay promotes to fp32, but v stays fp16 -> CompilationError.
Match _fwd_diag_kernel's pattern: explicit .to(tl.float32) on load.

+++ fastdeploy/model_executor/ops/triton_ops/lightning_attn.py
@@ -213,12 +213,12 @@ def _fwd_kv_parallel(
             K_trans_block_ptr - left_shift * d,
             mask=kv_index[None, :] >= left_bound,
             other=0.0,
-        )
+        ).to(tl.float32)
         v = tl.load(
             V_block_ptr - left_shift * e,
             mask=kv_index[:, None] >= left_bound,
             other=0.0,
-        )
+        ).to(tl.float32)
 
         # Load decay factor and compute weighted key-value outer product
         k_decay = tl.load(k_decay_ptr)
